+++ elk-client/src/main/java/org/acumos/elk/client/service/SnapshotServiceImpl.java
@@ -32,6 +32,8 @@ import java.util.List;
 import org.acumos.elk.client.transport.CreateSnapshot;
 import org.acumos.elk.client.transport.DeleteSnapshot;
 import org.acumos.elk.client.transport.ELkRepositoryMetaData;
+import org.acumos.elk.client.transport.ElasticStackIndiceResponse;
+import org.acumos.elk.client.transport.ElasticStackIndices;
 import org.acumos.elk.client.transport.ElasticsearchSnapshotsResponse;
 import org.acumos.elk.client.transport.ElkCreateSnapshotRequest;
 import org.acumos.elk.client.transport.ElkDeleteSnapshotRequest;
@@ -51,9 +53,16 @@ import org.elasticsearch.action.admin.cluster.snapshots.get.GetSnapshotsRequest;
 import org.elasticsearch.action.admin.cluster.snapshots.get.GetSnapshotsResponse;
 import org.elasticsearch.action.admin.cluster.snapshots.restore.RestoreSnapshotRequest;
 import org.elasticsearch.action.admin.cluster.snapshots.restore.RestoreSnapshotResponse;
+import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
+import org.elasticsearch.action.admin.indices.settings.get.GetSettingsRequest;
+import org.elasticsearch.action.admin.indices.settings.get.GetSettingsResponse;
+import org.elasticsearch.action.support.IndicesOptions;
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
+import org.elasticsearch.client.IndicesClient;
 import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestHighLevelClient;
+import org.elasticsearch.common.collect.ImmutableOpenMap;
+import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.snapshots.SnapshotInfo;
@@ -61,6 +70,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.carrotsearch.hppc.cursors.ObjectCursor;
 
 /**
  * Implementation of operation related to elastic stack snapshot. 
@@ -70,13 +81,12 @@ import org.springframework.stereotype.Service;
 public class SnapshotServiceImpl  extends AbstractELKClientConnection implements ISnapshotService {
 
 	private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-	
+
 	@Autowired
 	SnapshotRepositoryServiceImpl snapshotRepositoryServiceImpl;
 	
 	@Override
 	public ElkSnapshotsResponse getAllElasticSearchSnapshot(ElkRepositoriesRequest elkRepositoriesRequest) {
-
 		logger.debug("inside getAllElasticSearchSnapshot method ");
 		ElkSnapshotsResponse elkSnapshotsResponse = new ElkSnapshotsResponse();
 		List<ElasticsearchSnapshotsResponse> elasticsearchSnapshotsResponseList = new ArrayList<>();
@@ -87,51 +97,54 @@ public class SnapshotServiceImpl  extends AbstractELKClientConnection implements
 					elkRepositoriesRequest, eLkRepositoryMetaData.getName());
 			elasticsearchSnapshotsResponse.setRepositoryName(eLkRepositoryMetaData.getName());
 			elasticsearchSnapshotsResponseList.add(elasticsearchSnapshotsResponse);
-
 		}
 		elkSnapshotsResponse.setElasticsearchSnapshots(elasticsearchSnapshotsResponseList);
-
 		return elkSnapshotsResponse;
 	}
 
 	@Override
-	public ElkSnapshotsResponse createElasticSearchSnapshot(ElkCreateSnapshotRequest createDeleteSnapshotRequest)
+	public ElkSnapshotsResponse createElasticSearchSnapshot(ElkCreateSnapshotRequest createSnapshotUserRequest)
 			throws Exception {
-
 		logger.debug("Inside createElasticSearchSnapshot service method");
 		ElkSnapshotsResponse elkSnapshotsResponse = new ElkSnapshotsResponse();
-		if (createDeleteSnapshotRequest.getCreateSnapshots() != null
-				&& !createDeleteSnapshotRequest.getCreateSnapshots().isEmpty()) {
+		if (createSnapshotUserRequest.getCreateSnapshots() != null
+				&& !createSnapshotUserRequest.getCreateSnapshots().isEmpty()) {
 			String snapshotName = null;
 			ElasticsearchSnapshotsResponse snapshotResponse = new ElasticsearchSnapshotsResponse();
-			List<CreateSnapshot> createSnapshots = createDeleteSnapshotRequest.getCreateSnapshots();
+			List<CreateSnapshot> createSnapshots = createSnapshotUserRequest.getCreateSnapshots();
 			for (CreateSnapshot createSnapshot : createSnapshots) {
 				CreateSnapshotRequest createSnapshotRequest = new CreateSnapshotRequest();
-				if (createDeleteSnapshotRequest.getNodeTimeout() != null
-						&& Long.parseLong(createDeleteSnapshotRequest.getNodeTimeout()) > 2) {
+				if (createSnapshotUserRequest.getNodeTimeout() != null
+						&& Long.parseLong(createSnapshotUserRequest.getNodeTimeout()) > 2) {
 					createSnapshotRequest.masterNodeTimeout(
-							TimeValue.timeValueMinutes(Long.parseLong(createDeleteSnapshotRequest.getNodeTimeout())));
+							TimeValue.timeValueMinutes(Long.parseLong(createSnapshotUserRequest.getNodeTimeout())));
 				} else {
 					createSnapshotRequest.masterNodeTimeout(ElkClientConstants.TIME_TWO_MINT_OUT);
 				}
 				createSnapshotRequest.repository(createSnapshot.getRepositoryName());
-				if (createSnapshot.getCreate() != null && createSnapshot.getCreate().equalsIgnoreCase(ElkClientConstants.TRUE)) {
-					LocalDateTime dateTime = LocalDateTime.ofInstant(new Date().toInstant(), ZoneId.systemDefault());
-					String dateWithTime = dateTime.toString().replaceAll(":", "-")
-							.substring(0, dateTime.toString().length() - 4).toLowerCase();
-					snapshotName = "snapshot-" + dateWithTime;
-					createSnapshotRequest.snapshot(snapshotName);
-				}
+				LocalDateTime dateTime = LocalDateTime.ofInstant(new Date().toInstant(), ZoneId.systemDefault());
+				String dateWithTime = dateTime.toString().replaceAll(":", "-")
+						.substring(0, dateTime.toString().length() - 4).toLowerCase();
+				snapshotName = "snapshot-" + dateWithTime;
+				createSnapshotRequest.snapshot(snapshotName);
 
 				RestHighLevelClient client = restHighLevelClientConnection();
 				CreateSnapshotResponse createSnapshotResponse;
 				try {
-					logger.debug("Before snapshot create");
+					logger.debug("createSnapshot getIndices size: {}", createSnapshot.getIndices().size());
+					if (createSnapshot.getIndices() != null && createSnapshot.getIndices().size() > 0) {
+						List<String> indices = createSnapshot.getIndices();
+						createSnapshotRequest.indices(indices);
+					} else {
+						ElasticStackIndices elkIndicesResponse = getAllElasticSearchIndices();
+						List<String> indices = elkIndicesResponse.getIndices();
+						createSnapshotRequest.indices(indices);
+					}
 					createSnapshotResponse = client.snapshot().create(createSnapshotRequest, RequestOptions.DEFAULT);
 				} catch (ElasticsearchException ex) {
 					logger.debug("ElasticsearchException:", ex);
 					throw ex;
-				}catch (Exception ex) {
+				} catch (Exception ex) {
 					logger.debug("Exception:", ex);
 					throw ex;
 				} finally {
@@ -142,14 +155,11 @@ public class SnapshotServiceImpl  extends AbstractELKClientConnection implements
 				if (createSnapshotResponse != null && createSnapshotResponse.getSnapshotInfo() != null) {
 					SnapshotInfo snapshotInfo = createSnapshotResponse.getSnapshotInfo();
 					RestStatus status = createSnapshotResponse.status();
-					logger.debug("\nSnapshotInfo \n snapshotId: {}", snapshotInfo.snapshotId());
-					logger.debug("status: {}", snapshotInfo.status());
-					logger.debug("state: {}", snapshotInfo.state());
-					logger.debug("startTime: {}", snapshotInfo.startTime());
-					logger.debug("endTime: {}", snapshotInfo.endTime());
-					logger.debug("successfulShards: {}", snapshotInfo.successfulShards());
-					logger.debug("RestStatus: " + status.getStatus());
-
+					logger.debug(
+							"\\nSnapshotInfo \n snapshotId: {} \n status: {} \n state: {} \n startTime: {} \n endTime: {} \n successfulShards: {} \n RestStatus: {}",
+							snapshotInfo.snapshotId(), snapshotInfo.status(), snapshotInfo.state(),
+							snapshotInfo.startTime(), convertTime(snapshotInfo.endTime()),
+							convertTime(snapshotInfo.successfulShards()), status.getStatus());
 					List<ElasticsearchSnapshotsResponse> elasticsearchSnapshotsResponseList = new ArrayList<>();
 					List<ElkGetSnapshotMetaData> snapshotMetaDatas = new ArrayList<>();
 					ElkGetSnapshotMetaData elkGetSnapshotMetaData = new ElkGetSnapshotMetaData();
@@ -162,7 +172,6 @@ public class SnapshotServiceImpl  extends AbstractELKClientConnection implements
 					snapshotResponse.setSnapshots(snapshotMetaDatas);
 					snapshotResponse.setRepositoryName(createSnapshot.getRepositoryName());
 					elasticsearchSnapshotsResponseList.add(snapshotResponse);
-
 					elkSnapshotsResponse.setElasticsearchSnapshots(elasticsearchSnapshotsResponseList);
 				} else {
 					List<ElasticsearchSnapshotsResponse> elasticsearchSnapshotsResponseList = new ArrayList<>();
@@ -175,45 +184,29 @@ public class SnapshotServiceImpl  extends AbstractELKClientConnection implements
 					snapshotResponse.setSnapshots(snapshotMetaDatas);
 					snapshotResponse.setRepositoryName(createSnapshot.getRepositoryName());
 					elasticsearchSnapshotsResponseList.add(snapshotResponse);
-
 					elkSnapshotsResponse.setElasticsearchSnapshots(elasticsearchSnapshotsResponseList);
 				}
 			}
-
 		}
-
 		logger.debug("Snapshot is created(true for created)");
 		return elkSnapshotsResponse;
 	}
 
-
 	@Override
 	public ElkSnapshotsResponse deleteElasticSearchSnapshot(ElkDeleteSnapshotRequest elkDeleteSnapshotRequest) {
-
 		logger.debug("Inside deleteElasticSearchSnapshot service method");
-
 		List<DeleteSnapshot> deleteSnapshots = elkDeleteSnapshotRequest.getDeleteSnapshots();
-
 		RestHighLevelClient client = restHighLevelClientConnection();
 		if (deleteSnapshots != null && !deleteSnapshots.isEmpty()) {
-
 			ElkSnapshotsResponse elkSnapshotsResponse = new ElkSnapshotsResponse();
 			List<ElasticsearchSnapshotsResponse> elasticsearchSnapshotsResponseList = new ArrayList<>();
 			for (DeleteSnapshot deleteSnapshot : deleteSnapshots) {
 
-				if (deleteSnapshot.getDelete() != null && deleteSnapshot.getDelete().equalsIgnoreCase("true")
-						&& deleteSnapshot.getSnapShotId() != null) {
-
+				if (deleteSnapshot.getSnapShotId() != null) {
 					DeleteSnapshotRequest deleteSnapshotRequest = new DeleteSnapshotRequest(
 							deleteSnapshot.getRepositoryName());
-
-					String[] snapshotNameId = deleteSnapshot.getSnapShotId().split("/");
-					String snapshotName = snapshotNameId[0];
-					String snapshotId = snapshotNameId[1];
-					logger.debug("snapshotName {}", snapshotName);
-					logger.debug("snapshotId {}",snapshotId);
-
-					deleteSnapshotRequest.snapshot(snapshotName);
+					logger.debug("snapshotId: {}", deleteSnapshot.getSnapShotId());
+					deleteSnapshotRequest.snapshot(deleteSnapshot.getSnapShotId());
 					deleteSnapshotRequest.masterNodeTimeout(ElkClientConstants.TIME_ONE_MINT_OUT);
 					AcknowledgedResponse deleteSnapshotAcknowledgedResponse;
 					try {
@@ -242,48 +235,106 @@ public class SnapshotServiceImpl  extends AbstractELKClientConnection implements
 		}
 
 		return null;
-	}	
-	
-	
+	}
+
 	@Override
-	public ElkSnapshotsResponse restoreElasticSearchSnapshot(ElkRestoreSnapshotRequest elkRestoreSnapshotRequest) throws IOException {
-
+	public ElasticStackIndiceResponse restoreElasticSearchSnapshot(ElkRestoreSnapshotRequest elkRestoreSnapshotRequest)
+			throws IOException {
 		logger.debug("Inside restoreElasticSearchSnapshot service ");
+		ElasticStackIndiceResponse elasticStackIndiceResponse = new ElasticStackIndiceResponse();
 		RestHighLevelClient client = restHighLevelClientConnection();
-		logger.debug("Inside restoreElasticSearchSnapshot service 1111");
 		for (RestoreSnapshot restoreSnapshot : elkRestoreSnapshotRequest.getRestoreSnapshots()) {
-
 			try {
 				RestoreSnapshotRequest restoreSnapshotRequest = new RestoreSnapshotRequest(
 						elkRestoreSnapshotRequest.getRepositoryName(), restoreSnapshot.getSnapshotName());
-				RestoreSnapshotResponse restoreSnapshotResponse = client.snapshot().restore(restoreSnapshotRequest, RequestOptions.DEFAULT);
-				logger.debug("Inside restoreElasticSearchSnapshot service 5555 {}", restoreSnapshotResponse);
-				logger.debug("Inside restoreElasticSearchSnapshot service 6666 {}", restoreSnapshotResponse.getRestoreInfo());
-				
+				RestoreSnapshotResponse restoreSnapshotResponse = client.snapshot().restore(restoreSnapshotRequest,
+						RequestOptions.DEFAULT);
+				elasticStackIndiceResponse.setMessage(
+						"ElasticStack Snapshot restore is in progress, depending size it will take some time");
+				elasticStackIndiceResponse.setStatus(ElkClientConstants.SUCCESS);
+				logger.debug("RestoreSnapshotResponse : {}", elasticStackIndiceResponse.getMessage());
 			} catch (IOException e) {
-				logger.error("IOException ",e);
-				throw e;
+				logger.error("IOException: ", e);
+				throw new ErrorTransport("Unable to connect Elasticserach");
 			}
 		}
-
-		return null;
+		return elasticStackIndiceResponse;
 	}
-	
+
+	@Override
+	public ElasticStackIndices getAllElasticSearchIndices() throws IOException {
+
+		logger.debug("Inside getAllElasticSearchIndices");
+		RestHighLevelClient client = restHighLevelClientConnection();
+
+		ElasticStackIndices elkIndicesResponse = new ElasticStackIndices();
+		List<String> indices = new ArrayList<>();
+		try {
+			IndicesClient indicesAdminClient = client.indices();
+			GetSettingsRequest request = new GetSettingsRequest();
+			request.includeDefaults(true);
+			request.indicesOptions(IndicesOptions.lenientExpandOpen());
+			GetSettingsResponse getSettingsResponse = indicesAdminClient.getSettings(request, RequestOptions.DEFAULT);
+			ImmutableOpenMap<String, Settings> indexMap = getSettingsResponse.getIndexToDefaultSettings();
+			for (ObjectCursor<String> key : indexMap.keys()) {
+				logger.debug("key.value: {}", key.value);
+				indices.add(key.value);
+			}
+			elkIndicesResponse.setIndices(indices);
+		} catch (IOException e1) {
+			logger.debug("IOException: {}", e1);
+			throw new ErrorTransport("Unable to connect Elasticserach");
+		}
+
+		return elkIndicesResponse;
+	}
+
+	public ElasticStackIndiceResponse deleteElasticSearchIndices(ElasticStackIndices elasticStackIndices)
+			throws IOException {
+		logger.debug("Inside deleteElasticSearchIndices");
+		String response = null;
+		ElasticStackIndiceResponse elasticStackIndiceResponse = new ElasticStackIndiceResponse();
+		RestHighLevelClient client = restHighLevelClientConnection();
+		try {
+			IndicesClient indicesAdminClient = client.indices();
+			List<String> indices = elasticStackIndices.getIndices();
+			for (String indiceName : indices) {
+				DeleteIndexRequest deleteIndexRequest = new DeleteIndexRequest();
+				deleteIndexRequest.indices(indiceName);
+				AcknowledgedResponse acknowledgedResponse = indicesAdminClient.delete(deleteIndexRequest,
+						RequestOptions.DEFAULT);
+				if (acknowledgedResponse.isAcknowledged()) {
+					elasticStackIndiceResponse
+							.setMessage("Indices delete is in progress, depending size it will take some time");
+					elasticStackIndiceResponse.setStatus(ElkClientConstants.SUCCESS);
+				} else {
+					elasticStackIndiceResponse.setMessage("Indices is not deleted yet");
+					elasticStackIndiceResponse.setStatus(ElkClientConstants.FAIL);
+				}
+				logger.debug("{} ,Indice Name: {}", elasticStackIndiceResponse.getMessage(), indiceName);
+			}
+		} catch (IOException e) {
+			logger.debug("IOException: {}", e);
+			throw new ErrorTransport("Unable to connect Elasticserach");
+		}
+		return elasticStackIndiceResponse;
+	}
+
 	private ElasticsearchSnapshotsResponse getElasticsearchSnapshotDetails(
 			ElkRepositoriesRequest elkRepositoriesRequest, String repositoryName) {
-		
+
 		RestHighLevelClient client = restHighLevelClientConnection();
 		GetSnapshotsRequest getSnapshotsRequest = new GetSnapshotsRequest();
 		if (elkRepositoriesRequest.getRepositoryName() != null
 				&& !elkRepositoriesRequest.getRepositoryName().isEmpty()) {
 			getSnapshotsRequest.repository(elkRepositoriesRequest.getRepositoryName());
-		}else {
+		} else {
 			getSnapshotsRequest.repository(repositoryName);
 		}
-			
+
 		String[] snapshots = { "_all" };
 		getSnapshotsRequest.snapshots(snapshots);
-		 getSnapshotsRequest.masterNodeTimeout(ElkClientConstants.TIME_ONE_MINT_OUT);
+		getSnapshotsRequest.masterNodeTimeout(ElkClientConstants.TIME_ONE_MINT_OUT);
 		getSnapshotsRequest.verbose(true);
 		getSnapshotsRequest.ignoreUnavailable(false);
 		GetSnapshotsResponse response;
@@ -299,30 +350,32 @@ public class SnapshotServiceImpl  extends AbstractELKClientConnection implements
 
 		List<ElkGetSnapshotMetaData> snapshotMetaDatas = new ArrayList<>();
 		for (SnapshotInfo snapshotInfo : response.getSnapshots()) {
-			logger.debug("\nsnapshotId: {}", snapshotInfo.snapshotId());
-			logger.debug("status: {}", snapshotInfo.status());
-			logger.debug("state: {}", snapshotInfo.state());
-			logger.debug("startTime: {}", snapshotInfo.startTime());
-			logger.debug("endTime: {}", snapshotInfo.endTime());
-			logger.debug("successfulShards: {}", snapshotInfo.successfulShards());
+			logger.debug(
+					"\nsnapshotId: {} \n status: {} \n state: {} \n startTime: {} \n endTime: {} \n successfulShards: {}",
+					snapshotInfo.snapshotId(), snapshotInfo.status(), snapshotInfo.state(), snapshotInfo.startTime(),
+					convertTime(snapshotInfo.endTime()), convertTime(snapshotInfo.successfulShards()));
 			ElkGetSnapshotMetaData elkGetSnapshotMetaData = new ElkGetSnapshotMetaData();
-			elkGetSnapshotMetaData.setSnapShotId(snapshotInfo.snapshotId().toString());
+			String[] snapShotName = snapshotInfo.snapshotId().toString().split("/");
+			elkGetSnapshotMetaData.setSnapShotId(snapShotName[0]);
 			elkGetSnapshotMetaData.setStatus(snapshotInfo.status().toString());
 			elkGetSnapshotMetaData.setState(snapshotInfo.state().toString());
 			elkGetSnapshotMetaData.setStartTime(convertTime(snapshotInfo.startTime()));
 			elkGetSnapshotMetaData.setEndTime(convertTime(snapshotInfo.endTime()));
+			List<String> indices = snapshotInfo.indices();
+			for (String indice : indices) {
+				logger.debug("indices: {}", indice);
+			}
+			elkGetSnapshotMetaData.setIndices(indices);
 			snapshotMetaDatas.add(elkGetSnapshotMetaData);
 		}
 		snapshotResponse.setSnapshots(snapshotMetaDatas);
 		return snapshotResponse;
 	}
 
-	
-	public String convertTime(long time){
-	    Date date = new Date(time);
-	    Format format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-	    return format.format(date);
+	public String convertTime(long time) {
+		Date date = new Date(time);
+		Format format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+		return format.format(date);
 	}
-
 
 }